# docs/swagger.yaml
definitions:
  bouncer.Response:
    properties:
      allow:
        type: boolean
      callerid:
        type: string
      destination:
        type: string
    type: object
  handler.AuthorizationRequest:
    properties:
      endpoint:
        type: string
      extension:
        type: string
    type: object
  model.Endpoint:
    properties:
      codecs:
        items:
          type: string
        type: array
      context:
        type: string
      displayName:
        type: string
      extension:
        type: string
      id:
        type: string
      maxContacts:
        type: integer
      sid:
        type: integer
      transport:
        type: string
    type: object
  model.EndpointPage:
    properties:
      endpoints:
        items:
          $ref: '#/definitions/model.EndpointPageEntry'
        type: array
      retrieved:
        type: integer
      total:
        type: integer
    type: object
  model.EndpointPageEntry:
    properties:
      context:
        type: string
      displayName:
        type: string
      extension:
        type: string
      id:
        type: string
      sid:
        type: integer
    type: object
  model.NewEndpoint:
    properties:
      codecs:
        items:
          type: string
        type: array
      context:
        type: string
      displayName:
        type: string
      extension:
        type: string
      id:
        type: string
      maxContacts:
        type: integer
      password:
        type: string
      transport:
        type: string
    type: object
  model.PatchedEndpoint:
    properties:
      codecs:
        items:
          type: string
        type: array
      context:
        type: string
      displayName:
        type: string
      extension:
        type: string
      maxContacts:
        type: integer
      password:
        type: string
      transport:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: API to perform configuration management on Asterisk servers.
  title: Asterisk Administration API
  version: "1.0"
paths:
  /bouncer:
    post:
      consumes:
      - application/json
      parameters:
      - description: Action to be reviewed
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/handler.AuthorizationRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/bouncer.Response'
        "400":
          description: Bad Request
        "500":
          description: Internal Server Error
      summary: Determine whether the specified action (call) is allowed or not and
        provide details on how
      tags:
      - bouncer
  /endpoints:
    get:
      parameters:
      - default: 0
        description: Zero based page to fetch
        in: query
        name: page
        type: integer
      - default: 10
        description: Max amount of results to be returned
        in: query
        name: pageSize
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.EndpointPage'
        "400":
          description: Bad Request
        "500":
          description: Internal Server Error
      summary: List existing endpoints.
      tags:
      - endpoints
    post:
      consumes:
      - application/json
      parameters:
      - description: Endpoint's information
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/model.NewEndpoint'
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.Endpoint'
        "400":
          description: Bad Request
        "500":
          description: Internal Server Error
      summary: Create a new endpoint.
      tags:
      - endpoints
  /endpoints/{sid}:
    delete:
      parameters:
      - description: Sid of the endpoint to be deleted
        in: path
        name: sid
        required: true
        type: integer
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
        "500":
          description: Internal Server Error
      summary: Delete an endpoint and its associated resources.
      tags:
      - endpoints
    get:
      parameters:
      - description: Requested endpoint's sid
        in: path
        name: sid
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Endpoint'
        "400":
          description: Bad Request
        "500":
          description: Internal Server Error
      summary: Get information from a specific endpoint.
      tags:
      - endpoints
    patch:
      parameters:
      - description: Sid of the endpoint to be updated
        in: path
        name: sid
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.PatchedEndpoint'
        "400":
          description: Bad Request
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      summary: Update the specified endpoint. Omitted or null fields will remain unchanged.
      tags:
      - endpoints
swagger: "2.0"